moved to anonymous namespace and to DEFINE_DEVICE_TYPE_PRIVATE
